// job-portal-backend/src/common/enums/roles.enum.ts
export enum Role {
  ORG_SUPER_ADMIN = 'ORG_SUPER_ADMIN',
  ORG_ADMIN = 'ORG_ADMIN',
  RECRUITER = 'RECRUITER',
  PLATFORM_SUPER_ADMIN = 'PLATFORM_SUPER_ADMIN',
  PLATFORM_ADMIN = 'PLATFORM_ADMIN',
  CANDIDATE = 'CANDIDATE',
}

export enum RecruiterType {
  HR_RECRUITER = 'HR_RECRUITER',
  HIRING_MANAGER = 'HIRING_MANAGER',
  TECH_RECRUITER = 'TECH_RECRUITER',
}